% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDistinguisher.R
\name{gecd_CellDistinguisher}
\alias{gecd_CellDistinguisher}
\title{The gecd_CellDistinguisher function}
\usage{
gecd_CellDistinguisher(exprLinear, genesymb = NULL, numCellClasses = 2,
  minDistinguisherAlternatives = 1, maxDistinguisherAlternatives = 5,
  minAlternativesLengthsNormalized = 0.7, expressionQuantileForScale = 0.75,
  expressionQuantileForFilter = 1, expressionConcentrationRatio = 0.333,
  probesWithGenesOnly = FALSE, verbose = 0)
}
\arguments{
\item{exprLinear}{exprLinear is a matrix of (linear / non-logarithmic) expression values that has one row for each marker (a.k.a. probe or gene) and one column for each sample.}

\item{genesymb}{genesymb is the associated gene names for the rows of exprLinear. They are not passed in as names(exprLinear) because multiple rows (probes) may map to the same gene.}

\item{numCellClasses}{numCellClasses is the number of cell classes for which the function will locate class-specific distinguishers. Default is 2.}

\item{minDistinguisherAlternatives}{minDistinguisherAlternatives is the minimum number of distinguishers to be located for each cell class. Default is 1.}

\item{maxDistinguisherAlternatives}{maxDistinguisherAlternatives is the maximum number of distinguishers to be located for each cell class. Default is 5.}

\item{minAlternativesLengthsNormalized}{minAlternativesLengthsNormalized is the threshold for bestLengthsNormalized values. Only alternatives with a value at least as large as the threshold are retained. Default is 0.7. 0.0 is no filtering.}

\item{expressionQuantileForScale}{expressionQuantileForScale is in [0, 1]. Low values (e.g., 0.40) favor distinguishers for a cell class that have very little expression in the other cell classes. High values (e.g., 0.80) shift the balance towards distinguishers with high expression values. Default is 0.75. 0.0 is no preference for highly expressed genes.}

\item{expressionQuantileForFilter}{Default is 1.0. 1.0 is no filtering. 0.999 is some filtering.}

\item{expressionConcentrationRatio}{Default is 0.333. 0.0 is no filtering.}

\item{probesWithGenesOnly}{Default is FALSE.}

\item{verbose}{verbose indicates amount of printed output. Default is 0.}
}
\description{
Finds distinguishers that identify specific cell classes without having to know the cell classes.
}
\examples{
See https://github.com/GeneralElectric/CellDistinguisher/blob/master/README.txt
}
