% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDistinguisher.R
\name{gecd_DeconvolutionByDistinguishers}
\alias{gecd_DeconvolutionByDistinguishers}
\title{The gecd_DeconvolutionByDistinguishers function}
\usage{
gecd_DeconvolutionByDistinguishers(exprLinear, bestDistinguishers,
  nonNegativeOnly = TRUE, convexSolution = TRUE, verbose = 0)
}
\arguments{
\item{exprLinear}{exprLinear is the matrix to be factored}

\item{bestDistinguishers}{bestDistinguishers is the matrix of anchors that enable the factorization.}

\item{nonNegativeOnly}{If set to TRUE strictly enforces that the entries in the matrix factors be non-negative. Default is TRUE.}

\item{convexSolution}{If set to TRUE strictly enforces that the entries for a sample in the sampleComposition matrix sum to 1.0. Default is TRUE.}

\item{verbose}{Default is 0.}
}
\description{
Uses the distinguishers discovered by gecd_CellDistinguisher to deconvolve the original expression data.
}
\examples{
See https://github.com/GeneralElectric/CellDistinguisher/blob/master/README.txt
}
