% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDistinguisher.R
\name{gecd_DeconvolutionCellMix}
\alias{gecd_DeconvolutionCellMix}
\title{The gecd_DeconvolutionCellMix function}
\usage{
gecd_DeconvolutionCellMix(exprLinear, bestDistinguishers,
  nonNegativeOnly = TRUE, verbose = 0, log = FALSE, ...)
}
\arguments{
\item{exprLinear}{exprLinear is the matrix to be factored}

\item{bestDistinguishers}{bestDistinguishers is the matrix of anchors that enable the factorization.}

\item{nonNegativeOnly}{If set to TRUE strictly enforces that the entries in the matrix factors be non-negative. Default is TRUE.}

\item{verbose}{Default is 0.}

\item{log}{log parameter to be passed to the CellMix deconvolution tools. Default is FALSE.}

\item{...}{Additional parameters to be passed to the CellMix deconvolution tools.}
}
\description{
gecd_DeconvolutionCellMix is a wrapper for the CellMix deconvolution tools.  This is probably most useful with method="ssKL" and method="ssFrobenius".
}
\examples{
See https://github.com/GeneralElectric/CellDistinguisher/blob/master/README.txt
}
